Revert "Dex 22918 syntasa integration update"
