+++ 404.html
@@ -1,16 +1,14 @@
 <!DOCTYPE html><html><head><script type="importmap">{
   "imports": {
-    "@repobit/dex-constants": "https://cdn.jsdelivr.net/npm/@repobit/dex-constants@1.1.0/+esm",
-    "/npm/@repobit/dex-constants@1.1.0/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-constants@1.1.0/+esm",
-    "@repobit/dex-data-layer": "https://cdn.jsdelivr.net/npm/@repobit/dex-data-layer@1.6.1/+esm",
-    "/npm/@repobit/dex-data-layer@1.6.1/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-data-layer@1.6.1/+esm",
+    "@repobit/dex-data-layer": "https://cdn.jsdelivr.net/npm/@repobit/dex-data-layer@1.5.2/+esm",
+    "/npm/@repobit/dex-data-layer@1.5.2/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-data-layer@1.5.2/+esm",
     "@repobit/dex-store": "https://cdn.jsdelivr.net/npm/@repobit/dex-store@0.2.1/+esm",
     "/npm/@repobit/dex-store@0.2.1/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-store@0.2.1/+esm",
-    "@repobit/dex-utils": "https://cdn.jsdelivr.net/npm/@repobit/dex-utils@1.3.0/+esm",
-    "/npm/@repobit/dex-utils@1.3.0/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-utils@1.3.0/+esm",
+    "@repobit/dex-utils": "https://cdn.jsdelivr.net/npm/@repobit/dex-utils@1.2.1/+esm",
+    "/npm/@repobit/dex-utils@1.2.1/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-utils@1.2.1/+esm",
     "/npm/deepmerge@4.3.1/+esm": "https://cdn.jsdelivr.net//npm/deepmerge@4.3.1/+esm",
-    "@repobit/dex-launch": "https://cdn.jsdelivr.net/npm/@repobit/dex-launch@2.0.0/+esm",
-    "@repobit/dex-target": "https://cdn.jsdelivr.net/npm/@repobit/dex-target@2.0.1/+esm",
+    "@repobit/dex-launch": "https://cdn.jsdelivr.net/npm/@repobit/dex-launch@1.4.1/+esm",
+    "@repobit/dex-target": "https://cdn.jsdelivr.net/npm/@repobit/dex-target@1.9.1/+esm",
     "/npm/js-cookie@3.0.5/+esm": "https://cdn.jsdelivr.net//npm/js-cookie@3.0.5/+esm",
     "@tsparticles/engine": "https://cdn.jsdelivr.net/npm/@tsparticles/engine@3.8.1/+esm"
   }
