+++ _src/blocks/exit-popup/exit-popup.js
@@ -1,4 +1,4 @@
-import { target } from '../../scripts/target.js';
+import Target from '@repobit/dex-target';
 import { ProductInfo, Store } from '../../scripts/libs/store/store.js';
 // eslint-disable-next-line no-unused-vars
 export default async function decorate(block) {
@@ -8,7 +8,7 @@ export default async function decorate(block) {
   if (product) {
     const [alias, devices, years] = product.split(',');
     // eslint-disable-next-line no-undef
-    const exitPopupPromotion = (await target.getOffers({ mboxNames: 'exitPopupPid-mbox' }))?.promotion;
+    const exitPopupPromotion = (await Target.getOffers({ mboxNames: 'exitPopupPid-mbox' }))?.promotion;
     const products = await Store.getProducts([new ProductInfo(alias, 'consumer', exitPopupPromotion || custompid, true)]);
     const productItem = products[alias];
     const productCurrency = productItem.currency;
