+++ _src/blocks/footer/footer.js
@@ -1,7 +1,6 @@
 import { decorateIcons, getMetadata, loadBlocks } from '../../scripts/lib-franklin.js';
-import { getDomain } from '../../scripts/utils/utils.js';
+import { adobeMcAppendVisitorId, getDomain } from '../../scripts/utils/utils.js';
 import { decorateMain } from '../../scripts/scripts.js';
-import { adobeMcAppendVisitorId } from '../../scripts/target.js';
 
 function wrapImgsInLinks(container) {
   const pictures = container.querySelectorAll('picture');
