+++ _src/blocks/header/header.js
@@ -1,12 +1,12 @@
+import Target from '@repobit/dex-target';
 import { User, Cookies } from '@repobit/dex-utils';
 import {
   getMetadata, decorateIcons, decorateButtons, decorateTags,
 } from '../../scripts/lib-franklin.js';
 
 import {
-  getDomain, decorateBlockWithRegionId, decorateLinkWithLinkTrackingId,
+  adobeMcAppendVisitorId, getDomain, decorateBlockWithRegionId, decorateLinkWithLinkTrackingId,
 } from '../../scripts/utils/utils.js';
-import { adobeMcAppendVisitorId, target } from '../../scripts/target.js';
 
 import { Constants } from '../../scripts/libs/constants.js';
 
@@ -56,7 +56,7 @@ const updateMegaMenu = (username, email, newMegaMenuLoginTab) => {
   if (!userLoggedInExpirationDate
     || (userLoggedInExpirationDate && userLoggedInExpirationDate > Date.now())) {
     loginPopupLinksThatNeedToChange.forEach(async (loginPopupLink) => {
-      loginPopupLink.href = await target.appendVisitorIDsTo(loginPopupLink.dataset.loggedInLink);
+      loginPopupLink.href = await Target.appendVisitorIDsTo(loginPopupLink.dataset.loggedInLink);
     });
   }
 };
