+++ _src/scripts/lib-franklin.js
@@ -905,35 +905,10 @@ export function setup() {
   }
 }
 
-function initialiseSentry() {
-  window.sentryOnLoad = () => {
-    window.Sentry.init({
-      // Alternatively, use `process.env.npm_package_version` for a dynamic release version
-      // if your build tool supports it.
-      release: 'www-websites@1.0.0',
-      // Set tracesSampleRate to 1.0 to capture 100%
-      // of transactions for tracing.
-      // We recommend adjusting this value in production
-      // Learn more at
-      // https://docs.sentry.io/platforms/javascript/configuration/options/#traces-sample-rate
-      tracesSampleRate: 0.05,
-      // Capture Replay for 10% of all sessions,
-      // plus for 100% of sessions with an error
-      // Learn more at
-      // https://docs.sentry.io/platforms/javascript/session-replay/configuration/#general-integration-configuration
-      replaysSessionSampleRate: 0.1,
-      replaysOnErrorSampleRate: 1.0,
-
-      allowUrls: ['www.bitdefender.com'],
-    });
-  };
-}
-
 /**
  * Auto initializiation.
  */
 function init() {
-  initialiseSentry();
   setup();
   sampleRUM('top');
