+++ _src/scripts/target.js
@@ -1,26 +1,5 @@
 /* eslint-disable max-classes-per-file */
-import Target from '@repobit/dex-target';
-import { PageLoadStartedEvent } from '@repobit/dex-data-layer';
-import { User } from '@repobit/dex-utils';
 import { sampleRUM } from './lib-franklin.js';
-import page from './page.js';
-import {
-  getMetadata,
-  shouldABTestsBeDisabled,
-  generatePageLoadStartedName,
-  GLOBAL_EVENTS,
-} from './utils/utils.js';
-
-export const target = new Target({
-  pageLoadStartedEvent: new PageLoadStartedEvent(
-    page,
-    {
-      name: generatePageLoadStartedName(),
-      geoRegion: await User.country,
-      serverName: 'hlx.live', // indicator for AEM Success Edge
-    },
-  ),
-});
 
 /**
  * Convert a URL to a relative URL.
@@ -82,78 +61,5 @@ export async function runTargetExperiment(experimentUrl, experimentId) {
     };
   } catch (e) {
     return null;
-  }
-}
-
-export function appendAdobeMcLinks(selector) {
-  try {
-    const wrapperSelector = typeof selector === 'string' ? document.querySelector(selector) : selector;
-
-    const hrefSelector = '[href*=".bitdefender."]';
-    wrapperSelector.querySelectorAll(hrefSelector).forEach(async (link) => {
-      const destinationURLWithVisitorIDs = await target.appendVisitorIDsTo(link.href);
-      link.href = destinationURLWithVisitorIDs.replace(/MCAID%3D.*%7CMCORGID/, 'MCAID%3D%7CMCORGID');
-    });
-  } catch (e) {
-    // eslint-disable-next-line no-console
-    console.error(e);
-  }
-}
-
-/**
-* get experiment details from Target
-* @returns {Promise<{
-*  experimentId: string;
-*  experimentVariant: string;
-* } | null>}
-  */
-export const getTargetExperimentDetails = async () => {
-  /**
-   * @type {{
-   *  experimentId: string;
-   *  experimentVariant: string;
-   * }|null}
-   */
-  let targetExperimentDetails = null;
-
-  async function loadCSS(href) {
-    return new Promise((resolve, reject) => {
-      if (!document.querySelector(`head > link[href="${href}"]`)) {
-        const link = document.createElement('link');
-        link.rel = 'stylesheet';
-        link.href = href;
-        link.onload = resolve;
-        link.onerror = reject;
-        document.head.append(link);
-      } else {
-        resolve();
-      }
-    });
-  }
-
-  const targetExperimentLocation = getMetadata('target-experiment-location');
-  const targetExperimentId = getMetadata('target-experiment');
-  if (targetExperimentLocation && targetExperimentId && !shouldABTestsBeDisabled()) {
-    const offer = await target.getOffers({ mboxNames: targetExperimentLocation });
-    const { url, template } = offer || {};
-    if (template) {
-      loadCSS(`${window.hlx.codeBasePath}/scripts/template-factories/${template}.css`);
-      document.body.classList.add(template);
-    }
-    targetExperimentDetails = await runTargetExperiment(url, targetExperimentId);
-  }
-
-  return targetExperimentDetails;
-};
-
-export function adobeMcAppendVisitorId(selector) {
-  // https://experienceleague.adobe.com/docs/id-service/using/id-service-api/methods/appendvisitorid.html?lang=en
-
-  if (window.ADOBE_MC_EVENT_LOADED) {
-    appendAdobeMcLinks(selector);
-  } else {
-    document.addEventListener(GLOBAL_EVENTS.ADOBE_MC_LOADED, () => {
-      appendAdobeMcLinks(selector);
-    });
   }
 }
