+++ head.html
@@ -1,26 +1,23 @@
 <script type="importmap">{
   "imports": {
-    "@repobit/dex-constants": "https://cdn.jsdelivr.net/npm/@repobit/dex-constants@1.1.0/+esm",
-    "/npm/@repobit/dex-constants@1.1.0/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-constants@1.1.0/+esm",
-    "@repobit/dex-data-layer": "https://cdn.jsdelivr.net/npm/@repobit/dex-data-layer@1.6.1/+esm",
-    "/npm/@repobit/dex-data-layer@1.6.1/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-data-layer@1.6.1/+esm",
+    "@repobit/dex-data-layer": "https://cdn.jsdelivr.net/npm/@repobit/dex-data-layer@1.5.2/+esm",
+    "/npm/@repobit/dex-data-layer@1.5.2/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-data-layer@1.5.2/+esm",
     "@repobit/dex-store": "https://cdn.jsdelivr.net/npm/@repobit/dex-store@0.2.1/+esm",
     "/npm/@repobit/dex-store@0.2.1/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-store@0.2.1/+esm",
-    "@repobit/dex-utils": "https://cdn.jsdelivr.net/npm/@repobit/dex-utils@1.3.0/+esm",
-    "/npm/@repobit/dex-utils@1.3.0/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-utils@1.3.0/+esm",
+    "@repobit/dex-utils": "https://cdn.jsdelivr.net/npm/@repobit/dex-utils@1.2.1/+esm",
+    "/npm/@repobit/dex-utils@1.2.1/+esm": "https://cdn.jsdelivr.net//npm/@repobit/dex-utils@1.2.1/+esm",
     "/npm/deepmerge@4.3.1/+esm": "https://cdn.jsdelivr.net//npm/deepmerge@4.3.1/+esm",
-    "@repobit/dex-launch": "https://cdn.jsdelivr.net/npm/@repobit/dex-launch@2.0.0/+esm",
-    "@repobit/dex-target": "https://cdn.jsdelivr.net/npm/@repobit/dex-target@2.0.1/+esm",
+    "@repobit/dex-launch": "https://cdn.jsdelivr.net/npm/@repobit/dex-launch@1.4.1/+esm",
+    "@repobit/dex-target": "https://cdn.jsdelivr.net/npm/@repobit/dex-target@1.9.1/+esm",
     "/npm/js-cookie@3.0.5/+esm": "https://cdn.jsdelivr.net//npm/js-cookie@3.0.5/+esm",
     "@tsparticles/engine": "https://cdn.jsdelivr.net/npm/@tsparticles/engine@3.8.1/+esm"
   }
 }</script>
 <meta name="referrer" content="no-referrer-when-downgrade" />
 <meta name="viewport" content="width=device-width, initial-scale=1" />
-<script src="/_src/scripts/lib-franklin.js" type="module" defer=""></script>
-<script src="https://js.sentry-cdn.com/453d79512df247d7983074696546ca60.min.js" defer="" crossorigin="anonymous"></script>
-<script src="/_src/scripts/utils/utils.js" type="module" defer=""></script>
-<script src="/_src/scripts/scripts.js" type="module" defer=""></script>
+<script src="/_src/scripts/lib-franklin.js" type="module"></script>
+<script src="/_src/scripts/utils/utils.js" type="module"></script>
+<script src="/_src/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/_src/styles/styles.css" />
 <link rel="shortcut icon" href="/content/dam/bitdefender/favicon/favicon.ico" type="image/x-icon" />
 <link rel="icon" href="/content/dam/bitdefender/favicon/favicon.ico" type="image/x-icon" />
