+++ package.json
@@ -31,7 +31,7 @@
     "eslint": "8.35.0",
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-plugin-import": "2.27.5",
-    "happy-dom": "^17.4.7",
+    "happy-dom": "^17.4.4",
     "husky": "^9.1.7",
     "jsdom": "^25.0.1",
     "sinon": "15.0.1",
@@ -39,12 +39,11 @@
     "stylelint-config-standard": "30.0.1"
   },
   "dependencies": {
-    "@repobit/dex-constants": "^1.1.0",
-    "@repobit/dex-data-layer": "^1.6.1",
-    "@repobit/dex-launch": "^2.0.0",
+    "@repobit/dex-data-layer": "^1.5.2",
+    "@repobit/dex-launch": "^1.4.1",
     "@repobit/dex-store": "^0.2.1",
-    "@repobit/dex-target": "^2.0.1",
-    "@repobit/dex-utils": "^1.3.0",
+    "@repobit/dex-target": "^1.9.1",
+    "@repobit/dex-utils": "^1.2.1",
     "@tsparticles/engine": "^3.8.1"
   },
   "volta": {
